create dedicated feature for pubsub
