+++ src/NServiceBus.Core/Routing/PublishSubscribeFeature.cs
@@ -1,0 +1,78 @@
+namespace NServiceBus.Features
+{
+    using Routing;
+    using Routing.MessageDrivenSubscriptions;
+    using Transport;
+    using Unicast.Messages;
+    using Unicast.Subscriptions.MessageDrivenSubscriptions;
+
+    class PublishSubscribeFeature : Feature
+    {
+        public PublishSubscribeFeature()
+        {
+            EnableByDefault();
+            Defaults(s =>
+            {
+                s.SetDefault<Publishers>(new Publishers());
+                s.SetDefault<ConfiguredPublishers>(new ConfiguredPublishers());
+
+                s.SetDefault<EndpointInstances>(new EndpointInstances());
+                s.SetDefault<DistributionPolicy>(new DistributionPolicy());
+            });
+        }
+
+        protected internal override void Setup(FeatureConfigurationContext context)
+        {
+            var transportInfrastructure = context.Settings.Get<TransportInfrastructure>();
+            var canReceive = !context.Settings.GetOrDefault<bool>("Endpoint.SendOnly");
+            var conventions = context.Settings.Get<Conventions>();
+            var enforceBestPractices = context.Settings.Get<bool>(BestPracticesFeature.EnforceBestPracticesSettingsKey);
+
+            var distributionPolicy = context.Settings.Get<DistributionPolicy>();
+            var endpointInstances = context.Settings.Get<EndpointInstances>();
+            var publishers = context.Settings.Get<Publishers>();
+            var configuredPublishers = context.Settings.Get<ConfiguredPublishers>();
+
+            configuredPublishers.Apply(publishers, conventions, enforceBestPractices);
+
+            var outboundRoutingPolicy = transportInfrastructure.OutboundRoutingPolicy;
+
+            // pub/sub
+            if (outboundRoutingPolicy.Publishes == OutboundRoutingType.Unicast)
+            {
+                context.Pipeline.Register(b =>
+                {
+                    var unicastPublishRouter = new UnicastPublishRouter(b.Build<MessageMetadataRegistry>(), b.Build<ISubscriptionStorage>());
+                    return new UnicastPublishRouterConnector(unicastPublishRouter, distributionPolicy);
+                }, "Determines how the published messages should be routed");
+            }
+            else
+            {
+                context.Pipeline.Register(new MulticastPublishRouterBehavior(), "Determines how the published messages should be routed");
+            }
+
+            if (canReceive)
+            {
+                var distributorAddress = context.Settings.GetOrDefault<string>("LegacyDistributor.Address");
+
+                // pub/sub
+                if (outboundRoutingPolicy.Publishes == OutboundRoutingType.Unicast)
+                {
+                    var subscriberAddress = distributorAddress ?? context.Settings.LocalAddress();
+                    var subscriptionRouter = new SubscriptionRouter(publishers, endpointInstances, i => transportInfrastructure.ToTransportAddress(LogicalAddress.CreateRemoteAddress(i)));
+
+                    context.Pipeline.Register(b => new MessageDrivenSubscribeTerminator(subscriptionRouter, subscriberAddress, context.Settings.EndpointName(), b.Build<IDispatchMessages>()), "Sends subscription requests when message driven subscriptions is in use");
+                    context.Pipeline.Register(b => new MessageDrivenUnsubscribeTerminator(subscriptionRouter, subscriberAddress, context.Settings.EndpointName(), b.Build<IDispatchMessages>()), "Sends requests to unsubscribe when message driven subscriptions is in use");
+                }
+                else
+                {
+                    var transportSubscriptionInfrastructure = transportInfrastructure.ConfigureSubscriptionInfrastructure();
+                    var subscriptionManager = transportSubscriptionInfrastructure.SubscriptionManagerFactory();
+
+                    context.Pipeline.Register(new NativeSubscribeTerminator(subscriptionManager), "Requests the transport to subscribe to a given message type");
+                    context.Pipeline.Register(new NativeUnsubscribeTerminator(subscriptionManager), "Requests the transport to unsubscribe to a given message type");
+                }
+            }
+        }
+    }
+}
